requirements.txt
customtkinter
Pillow
numpy
tkinterdnd2
